+++ python/ray/rllib/rollout.py
@@ -12,7 +12,6 @@ import pickle
 import gym
 import ray
 from ray.rllib.agents.registry import get_agent_class
-from ray.rllib.evaluation.sampler import clip_action
 from ray.tune.util import merge_dicts
 
 EXAMPLE_USAGE = """
@@ -154,11 +153,7 @@ def rollout(agent, env_name, num_steps, out=None, no_render=True):
                 else:
                     action = agent.compute_action(state)
 
-            if agent.config["clip_actions"]:
-                clipped_action = clip_action(action, env.action_space)
-                next_state, reward, done, _ = env.step(clipped_action)
-            else:
-                next_state, reward, done, _ = env.step(action)
+            next_state, reward, done, _ = env.step(action)
 
             if multiagent:
                 done = done["__all__"]
